pyproject.toml
[project]
name = "azx"
version = "0.1.4"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markdown-it-py>=3.0.0",
    "openai>=1.99.6",
    "prompt-toolkit>=3.0.51",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "socksio>=1.0.0",
]

[project.scripts]
azx = "azx:main"
